backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
psycopg[binary]==3.2.6
dj-database-url==2.3.0
celery[redis]==5.4.0
redis==5.2.1
django-redis==5.4.0
django-cors-headers==4.7.0
PyPDF2==3.0.1
python-docx==1.1.2
gunicorn==23.0.0
